+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fixes bug with incorrect use of one-time bindings, resulting in the
   dates not updating.
 
+- Fixes bug with empty weeks being added, messing up the row heights.
+
 ## 0.2.7
 
 - Adds selection of multiple dates